+++ lib/generators/bower_rails/initialize/templates/bower_rails.rb
@@ -1,0 +1,6 @@
+BowerRails.configure do |bower_rails|
+  # Uncomment this line if you want `rake bower:resolve` task to be run
+  # automatically before the `rake assets:precompile` task.
+
+  # bower_rails.resolve_before_precompile = true
+end